1.Created connectionPool class 2.Created CouponClientFacade interface 3.Created AdminFacade class 4.Created CompanyFacade class 5.Created CustomerFacade class

+++ CouponSystem/src/DB/DAO/CouponDAO.java
@@ -15,7 +15,7 @@ public interface CouponDAO {
 
 	void updateCoupon(Coupon coupon) throws Exception;
 
-	Coupon getCoupon(int id) throws Exception;
+	Coupon getCoupon(long id) throws Exception;
 	
 	Set<Coupon> getAllCoupouns() throws Exception;
